Introduce libxayagamecmain for language interop.
The newly introduced libxayagamecmain defines a main() function for a
full game daemon.  For the game specifics (like the game ID and the
actual game logic), declared but not defined extern "C" functions are
called.

This makes it easy to implement game rules in languages other than C++
without the need to wrap libxayagame itself or write any C++:  The
provided main() function can be used, and only needs to be linked
together with code that defines the declared functions in the target
language's "C interop" mode.